Give article body a more consistent and better name structure
