+++ packages/article/article-body.js
@@ -1,73 +1,28 @@
 import React from "react";
 import PropTypes from "prop-types";
-import { renderTrees } from "@times-components/markup";
-import ArticleImage from "@times-components/article-image";
-
-import {
-  ParagraphContainer,
-  Paragraph,
-  PrimaryImg,
-  SecondaryImg,
-  InlineImg
-} from "./styles/body/responsive";
-import styles from "./styles/body";
-
-const getImageContainer = imageType => {
-  switch (imageType) {
-    case "secondary":
-      return SecondaryImg;
-    case "inline":
-      return InlineImg;
-    default:
-      return PrimaryImg;
-  }
-};
+import ArticleRow from "./article-body-row";
 
 const ArticleBody = props => {
-  const { data, index } = props.content;
-  return renderTrees([data], {
-    paragraph(key, attributes, children) {
-      return (
-        <ParagraphContainer
-          testID={`paragraph-${index}`}
-          accessibilityLabel={`paragraph-${index}`}
-          key={key}
-          style={[styles.articleMainContentRow]}
-        >
-          <Paragraph style={styles.articleTextElement}>{children}</Paragraph>
-        </ParagraphContainer>
-      );
-    },
-    image(key, attributes) {
-      const ImageContainer = getImageContainer(attributes.display);
-      return (
-        <ImageContainer key={key}>
-          <ArticleImage
-            imageOptions={{
-              display: attributes.display,
-              ratio: attributes.ratio,
-              url: attributes.url
-            }}
-            captionOptions={{
-              caption: attributes.caption,
-              credits: attributes.credits
-            }}
-          />
-        </ImageContainer>
-      );
-    }
-  });
+  const contentArray = props.content.map((data, index) => ({
+    data,
+    index
+  }));
+
+  const BodyView = contentArray.map(content => (
+    <ArticleRow key={`cont-${content.index}`} content={content} />
+  ));
+
+  return BodyView;
 };
 
 ArticleBody.propTypes = {
-  content: PropTypes.shape({
-    data: PropTypes.shape({
+  content: PropTypes.arrayOf(
+    PropTypes.shape({
       attributes: PropTypes.object,
       children: PropTypes.arrayOf(PropTypes.object),
       name: PropTypes.string
-    }),
-    index: PropTypes.number
-  }).isRequired
+    })
+  ).isRequired
 };
 
 export default ArticleBody;
